Module rmc_core::comms

The comms module contains the code for running the HTTP server

Structs

Drive

ServerSender
The ServerSender is a view into a RobotCommunicator that other threads/objects can use to request that messages be sent.

ServerState

Statics

static_rocket_route_info_for_files
Rocket code generated static route info.

static_rocket_route_info_for_handle_brake
Rocket code generated static route info.

static_rocket_route_info_for_handle_disable_drive
Rocket code generated static route info.

static_rocket_route_info_for_handle_drive
Rocket code generated static route info.

static_rocket_route_info_for_handle_enable_drive
Rocket code generated static route info.

static_rocket_route_info_for_handle_kill
Rocket code generated static route info.

static_rocket_route_info_for_handle_revive
Rocket code generated static route info.

static_rocket_route_info_for_index
Rocket code generated static route info.

Traits

SendableMessage
A SendableMessage is an object that can be encoded as a message and sent off to another device.

Functions

files

handle_brake

handle_disable_drive

handle_drive

handle_enable_drive

handle_kill

handle_revive

index

rocket_route_fn_files
Rocket code generated wrapping route function.

rocket_route_fn_handle_brake
Rocket code generated wrapping route function.

rocket_route_fn_handle_disable_drive
Rocket code generated wrapping route function.

rocket_route_fn_handle_drive
Rocket code generated wrapping route function.

rocket_route_fn_handle_enable_drive
Rocket code generated wrapping route function.

rocket_route_fn_handle_kill
Rocket code generated wrapping route function.

rocket_route_fn_handle_revive
Rocket code generated wrapping route function.

rocket_route_fn_index
Rocket code generated wrapping route function.

stage
Launches the server