Trait rmc_core::devices::motor_controllers::MotorController

pub trait MotorController: Send {
    fn set_speed(&mut self, new_speed: f32);

    fn stop(&mut self);

    fn invert(&mut self);

    fn get_motor_state(&self) -> MotorState;
}

Required methods

fn set_speed(&mut self, new_speed: f32)

Sets the current speed of the motor controller. The speed should be a floating point number between -1 and 1. A negative speed indicates that the direction is reversed.

fn stop(&mut self)

Sets the current speed of the motor controller to zero.

fn invert(&mut self)

Inverts the directionality of the motor controller.

fn get_motor_state(&self) -> MotorState

Returns the current motor state

Loading content...