Struct rmc_core::mechatronics::MechatronicsMessageSender

pub struct MechatronicsMessageSender {
    channel: Sender<MechatronicsCommand>,
    state: Arc<GlobalRobotState>,
}

Fields
channel: Sender<MechatronicsCommand>state: Arc<GlobalRobotState>
Methods

impl MechatronicsMessageSender

pub fn new(
channel: Sender<MechatronicsCommand>,
state: Arc<GlobalRobotState>
) -> Self

pub fn revive(&self)

pub fn kill(&self)

pub fn drive(&self, left: f32, right: f32) -> Result<(), ()>

Instructs the drive train to begin moving both sides at the provided speeds. Both speeds should be between -1 and 1. If any of the supplied speeds fall outside of that range, the command will not be sent and this method will return Err(LogData). A negative speed indicates that the motors should be run in reverse.

pub fn switch_to_drive(&self)

pub fn switch_to_dig(&self)

pub fn switch_to_dump(&self)

pub fn brake(&self)

Instructs the drive train to begin braking, halting all motion.

pub fn dump(&self)

pub fn reset_dumper(&self)

pub fn stop_dumper(&self)

pub fn dig(&self)

pub fn stop_digger(&self)

pub fn raise_ladder(&self)

pub fn lower_ladder(&self)

pub fn stop_actuators(&self)

fn send_command(&self, command: MechatronicsCommand)