Module rmc_core::comms

The comms module contains all code for controlling the communications infrastructure.

Modules

driver_station
Contains the implementation for communicating with the Driver Station. Also contains factory methods for instantiating that implementation.

io
Contains drivers for communicating with over different protocols or interfaces.

parsing
Contains code related to parsing.

Structs

CommsView
The CommsView is a view into a RobotCommunicator that other threads/objects can use to request that messages be sent.

RobotCommunicator
A RobotCommunicator is a complete Comms system which can be run. It has a CommsController that it owns and a driver that it owns for io purposes. The C type parameter is the type of the CommsController. The I type parameter is the type of the IoServerManager.

Traits

CommsController
A CommsController is an object with methods that are called to run the Comms by the RobotCommunicator.

SendableMessage
A SendableMessage is an object that can be encoded as a message and sent off to another device.

Functions

get_wrong_arg_count_log
Returns back a LogData for a message which had the wrong number of args.